requirements.txt
python-docx==1.1.2
lxml
beautifulsoup4
pandas
mysql
mysql-connector-python
requests